move getDocs to shared package for re-use
