+++ packages/shared/package.json
@@ -7,6 +7,7 @@
     "oas-bundle-size": "bin/bundle-size.mjs",
     "oas-lint": "bin/lint.sh"
   },
+  "main": "src/index.mjs",
   "scripts": {
     "lint": "bin/lint.sh"
   },
@@ -29,7 +30,9 @@
     "eslint-plugin-react": "^7.37.1",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-testing-library": "^6.3.0",
+    "fast-glob": "^3.3.2",
     "globals": "^15.10.0",
+    "react-docgen-typescript": "^2.2.2",
     "typescript-eslint": "^7.18.0"
   }
 }
